User/views.py
from rest_framework import generics # type:ignore
from .models import User
from .serializers import UserSerializer

class SignupListCreate(generics.ListCreateAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer

class SignupRetrieveUpdateDelete(generics.RetrieveUpdateDestroyAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer
