requirements.txt
django
djangorestframework 
drf-yasg
psycopg2-binary
django-cors-headers